add lombok, mapstruc libraries, standardlized api , remove autowired to using constructor

+++ demeOnlineShop/src/main/java/com/hoangmike/service/impl/UserServiceImpl.java
@@ -1,32 +1,37 @@
 package com.hoangmike.service.impl;
 
-import com.hoangmike.dto.UserCreationRequest;
+import com.hoangmike.dto.request.UserCreationRequest;
+import com.hoangmike.dto.response.UserResponse;
 import com.hoangmike.entity.User;
+import com.hoangmike.exception.AppException;
+import com.hoangmike.exception.ErrorCode;
+import com.hoangmike.mapper.UserMapper;
 import com.hoangmike.repository.UserRepository;
 import com.hoangmike.service.UserService;
-import org.springframework.beans.factory.annotation.Autowired;
+import lombok.AccessLevel;
+import lombok.RequiredArgsConstructor;
+import lombok.experimental.FieldDefaults;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 
 @Service
+@RequiredArgsConstructor
+@FieldDefaults(level = AccessLevel.PRIVATE, makeFinal = true)
 public class UserServiceImpl implements UserService {
-    @Autowired
-    private UserRepository userRepository;
+    UserRepository userRepository;
+    UserMapper userMapper;
 
     @Override
-    public User createUser(UserCreationRequest request) {
+    public UserResponse createUser(UserCreationRequest request) {
+
         User user = new User();
-        if(userRepository.existsByUserName(request.getUserName()))
-            throw new RuntimeException("Username already exists");
-        user.setUserName(request.getUserName());
-        user.setPassword(request.getPassword());
-        user.setEmail(request.getEmail());
-        user.setPhone(request.getPhone());
-        user.setFullName(request.getFullName());
-        user.setUserStatus(request.isUserStatus());
-        user.setRoleId(request.getRoleId());
-        return userRepository.save(user);
+        if(userRepository.existsByUserName(request.getUserName())) {
+            throw new AppException(ErrorCode.USER_EXISTED);
+        }
+        else userMapper.AddOrUpdateUser(user,request);
+
+        return userMapper.toUserResponse(userRepository.save(user));
     }
 
 
@@ -36,28 +41,27 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public User getUserById(String id) {
-        return userRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("User not found")); 
+    public UserResponse getUserById(String id) {
+
+        return userMapper.toUserResponse(userRepository.findById(id)
+                .orElseThrow(() -> new AppException(ErrorCode.USER_NOT_FOUND)));
     }
 
     @Override
-    public User updateUserById(String userId, UserCreationRequest request) {
-        User user = getUserById(userId);
+    public UserResponse updateUserById(String userId, UserCreationRequest request) {
+        User user = userRepository.findById(userId).orElseThrow(() -> new AppException(ErrorCode.USER_NOT_FOUND));
+
         if(userRepository.existsByUserName(request.getUserName()))
-            throw new RuntimeException("Username already exists");
-        user.setUserName(request.getUserName());
-        user.setPassword(request.getPassword());
-        user.setEmail(request.getEmail());
-        user.setPhone(request.getPhone());
-        user.setFullName(request.getFullName());
-        user.setUserStatus(request.isUserStatus());
-        user.setRoleId(request.getRoleId());
-        return userRepository.save(user);
+            throw new AppException(ErrorCode.USER_EXISTED);
+        else userMapper.AddOrUpdateUser(user, request);
+
+        return userMapper.toUserResponse(userRepository.save(user));
     }
 
     @Override
     public void deleteUserById(String id) {
-        userRepository.deleteById(id);
+        if(userRepository.existsById(id))
+            userRepository.deleteById(id);
+        else throw new AppException(ErrorCode.USER_NOT_FOUND);
     }
 }
